dashboard refactor phase 7d (#3473)
* extract shared components and hooks

* create ProgramEnrollmentCard and use it

* add EnrolledCourseCard component

* add UnenrolledCourseCard

* set up CoursewareCard to use type discrimination and call the proper isolated component, then use that everywhere

* roll ModuleCard style and behavior into shared components

* update OrganizationCards to use the SharedCard component

* update dashboarddialog tests to use new shared components

* delete dashboardcard and modulecard

* fix border radius on CTA

* proper default for coursewareUrl

* fix typo

* fix separator gating on compact layout

* fix certificate link regression

* fix enrollment status indicator

* restore GTM event

* add regression test

* remove dead code

* refactor(CoursewareCard): dispatch on `kind`, make `layout` density-only (#3478)

The router overloaded `layout` as both the union discriminant
(`"program"`/`"moduleRow"`) and a presentation knob (`"default"`/`"compact"`),
and re-derived enrolled-vs-unenrolled from `displayedEnrollment` in two arms.

Split the two axes:
- `kind` ("course" | "enrollment" | "program-enrollment") is the sole
  discriminant.
- `layout` ("default" | "compact") carries density only.

Consequences:
- Drops the `{ displayedEnrollment } & Partial<DashboardCourseEntry>` bridge
  type. Home uses a first-class `kind="enrollment"` arm instead of faking an
  entry; the `"course"` arm now takes a real `DashboardCourseEntry`.
- Collapses the duplicated enrolled/unenrolled dispatch. The `"enrollment"` and
  `"course"`-enrolled render blocks are kept deliberately separate (not shared):
  they are identical today but only the `"course"` arm will grow the future
  multi-run expand/collapse UI.
- Removes the dead `showNotComplete`/`offerUpgrade`/`isLoading`/
  `contextMenuItems`/`noun` props the router declared but never forwarded
  (the leaf cards derive these internally).

Callsites migrated: HomeEnrollmentsDisplay, ContractContent (also restores the
dropped `Component="li"`), ProgramAsCourseCard (`moduleRow` -> `compact`),
ProgramEnrollmentDisplay, and DashboardDialogs.test. The dialog-test fixtures
now build entries via `buildCourseEntry` rather than hand-rolled literals, so
`displayedRun` is the real enrollable run instead of a contradicted `null`.

Behavior-preserving: 471 dashboard tests pass, typecheck + lint clean.


* only use text variant when module is complete for compact module rows

* test enrollment redirect

* use B2BEnrollRequestRequest directly as an argument

* rename styled component

* delete the refactor plan

* use secondary style on compact row buttons that are unenrolled

---------
