--- frontends/main/src/app-pages/DashboardPage/CoursewareDisplay/DashboardCard.tsx
@@ -1,1057 +1,0 @@
-import React from "react"
-import {
-  styled,
-  Link,
-  SimpleMenu,
-  SimpleMenuItem,
-  Stack,
-  Skeleton,
-  LoadingSpinner,
-} from "ol-components"
-import NextLink from "next/link"
-import { ActionButton, Button, ButtonLink } from "@mitodl/smoot-design"
-import {
-  RiArrowRightLine,
-  RiAddLine,
-  RiMore2Line,
-  RiAwardLine,
-} from "@remixicon/react"
-import { calendarDaysUntil, isInPast, NoSSR } from "ol-utilities"
-import { useFeatureFlagEnabled } from "posthog-js/react"
-import { FeatureFlags } from "@/common/feature_flags"
-
-import { EnrollmentStatusIndicator } from "./EnrollmentStatusIndicator"
-import { getReceiptMenuItem } from "./receiptMenuItem"
-import {
-  EmailSettingsDialog,
-  JustInTimeDialog,
-  UnenrollDialog,
-  UnenrollProgramDialog,
-} from "./DashboardDialogs"
-import NiceModal from "@ebay/nice-modal-react"
-import {
-  useCreateB2bEnrollment,
-  useCreateEnrollment,
-  useCreateVerifiedProgramEnrollment,
-} from "api/mitxonline-hooks/enrollment"
-import { mitxUserQueries } from "api/mitxonline-hooks/user"
-import { useQuery } from "@tanstack/react-query"
-import { coursePageView, programPageView, programView } from "@/common/urls"
-import {
-  getCourseEnrollmentAction,
-  isVerifiedEnrollmentMode,
-  mitxonlineLegacyUrl,
-} from "@/common/mitxonline"
-import { useReplaceBasketItem } from "@/common/mitxonline/useReplaceBasketItem"
-import { EnrollmentStatus, getBestRun, getEnrollmentStatus } from "./helpers"
-import { trackCourseEnrolled } from "@/common/analytics/gtm"
-import {
-  CourseWithCourseRunsSerializerV2,
-  V3UserProgramEnrollment,
-  BaseCourseRun,
-  CourseRunV2,
-  DisplayModeEnum,
-} from "@mitodl/mitxonline-api-axios/v2"
-import CourseEnrollmentDialog from "@/page-components/EnrollmentDialogs/CourseEnrollmentDialog"
-import {
-  DashboardType,
-  type DashboardResource,
-} from "./model/dashboardViewModel"
-
-// Re-exported for consumers that still import from this file.
-export type { DashboardResource }
-export { DashboardType }
-
-/**
- * Gets the certificate link for a dashboard resource based on its type.
- */
-const getCertificateLink = (resource: DashboardResource): string | null => {
-  if (resource.type === DashboardType.CourseRunEnrollment) {
-    const link = resource.data.certificate?.link
-    if (!link) return null
-    const pattern = /\/certificate\/([^/]+)\/?$/
-    return link.replace(pattern, "/certificate/course/$1/")
-  }
-  if (resource.type === DashboardType.ProgramEnrollment) {
-    const link = resource.data.certificate?.link
-    if (!link) return null
-    const pattern = /\/certificate\/([^/]+)\/?$/
-    return link.replace(pattern, "/certificate/program/$1/")
-  }
-  return null
-}
-
-const CardRoot = styled.div<{
-  screenSize: "desktop" | "mobile"
-  variant?: "default" | "stacked"
-}>(({ theme, screenSize, variant = "default" }) => [
-  {
-    position: "relative",
-    border: `1px solid ${theme.custom.colors.lightGray2}`,
-    backgroundColor: theme.custom.colors.white,
-    padding: "16px",
-    display: "flex",
-    gap: "8px",
-    alignItems: "center",
-  },
-  // Mobile styles for default variant
-  variant === "default" && {
-    [theme.breakpoints.down("md")]: {
-      border: "none",
-      borderBottom: `1px solid ${theme.custom.colors.lightGray2}`,
-      borderRadius: "0px",
-      boxShadow: "none",
-      flexDirection: "column",
-      gap: "16px",
-    },
-  },
-  // Stacked variant styles
-  variant === "stacked" && {
-    border: "none",
-    borderBottom: `1px solid ${theme.custom.colors.lightGray2}`,
-    borderRadius: "0px !important",
-    boxShadow: "none",
-    "&:first-of-type": {
-      borderTopLeftRadius: "8px !important",
-      borderTopRightRadius: "8px !important",
-    },
-    "&:last-of-type": {
-      borderBottomLeftRadius: "8px !important",
-      borderBottomRightRadius: "8px !important",
-      borderBottom: "none",
-    },
-    [theme.breakpoints.down("md")]: {
-      flexDirection: "column",
-      gap: "16px",
-    },
-  },
-  screenSize === "desktop" && {
-    [theme.breakpoints.down("md")]: {
-      display: "none",
-    },
-  },
-  screenSize === "mobile" && {
-    [theme.breakpoints.up("md")]: {
-      display: "none",
-    },
-  },
-])
-
-const TitleHeading = styled.h3(({ theme }) => ({
-  margin: 0,
-  [theme.breakpoints.down("md")]: {
-    maxWidth: "calc(100% - 16px)",
-  },
-}))
-
-const TitleLink = styled(Link)()
-
-const TitleText = styled.h3<{ clickable?: boolean }>(
-  ({ theme, clickable }) => ({
-    margin: 0,
-    ...theme.typography.subtitle2,
-    color: theme.custom.colors.darkGray2,
-    cursor: clickable ? "pointer" : "default",
-    [theme.breakpoints.down("md")]: {
-      maxWidth: "calc(100% - 16px)",
-    },
-  }),
-)
-
-const MenuButton = styled(ActionButton)<{
-  status: EnrollmentStatus
-}>(({ theme, status }) => [
-  {
-    marginLeft: "-8px",
-    [theme.breakpoints.down("md")]: {
-      position: "absolute",
-      top: "0",
-      right: "0",
-    },
-  },
-  status !== EnrollmentStatus.Completed &&
-    status !== EnrollmentStatus.Enrolled && {
-      visibility: "hidden",
-    },
-])
-
-const getContextMenuItems = (
-  title: string,
-  resource: DashboardResource,
-  useProductPages: boolean,
-  additionalItems: SimpleMenuItem[] = [],
-  hideDetailsUrl = false,
-) => {
-  const menuItems = []
-  if (resource.type === DashboardType.ProgramEnrollment) {
-    const { program } = resource.data
-    const detailsUrl = useProductPages
-      ? programPageView({
-          readable_id: program.readable_id,
-          display_mode: program.display_mode,
-        })
-      : mitxonlineLegacyUrl(`/programs/${program.readable_id}`)
-
-    if (!hideDetailsUrl && detailsUrl) {
-      menuItems.push({
-        className: "dashboard-card-menu-item",
-        key: "view-program-details",
-        label: "View Program Details",
-        href: detailsUrl,
-      })
-    }
-
-    menuItems.push({
-      className: "dashboard-card-menu-item",
-      key: "program-record",
-      label: "Program Record",
-      href: mitxonlineLegacyUrl(`/records/${program.id}/`),
-    })
-
-    if (
-      program.display_mode !== DisplayModeEnum.Course &&
-      !isVerifiedEnrollmentMode(resource.data.enrollment_mode)
-    ) {
-      menuItems.push({
-        className: "dashboard-card-menu-item",
-        key: "unenroll-program",
-        label: "Unenroll",
-        onClick: () => {
-          NiceModal.show(UnenrollProgramDialog, {
-            title,
-            programId: program.id,
-          })
-        },
-      })
-    }
-
-    const receiptMenuItem = getReceiptMenuItem(
-      resource.data.enrollment_mode,
-      `/orders/receipt/by-program/${program.id}/`,
-    )
-    if (receiptMenuItem) menuItems.push(receiptMenuItem)
-  }
-  if (resource.type === DashboardType.CourseRunEnrollment) {
-    const detailsUrl = useProductPages
-      ? coursePageView(resource.data.run.course.readable_id)
-      : mitxonlineLegacyUrl(`/courses/${resource.data.run.course.readable_id}`)
-
-    const courseMenuItems = []
-
-    if (!hideDetailsUrl && detailsUrl) {
-      courseMenuItems.push({
-        className: "dashboard-card-menu-item",
-        key: "view-course-details",
-        label: "View Course Details",
-        href: detailsUrl,
-      })
-    }
-
-    courseMenuItems.push(
-      {
-        className: "dashboard-card-menu-item",
-        key: "email-settings",
-        label: "Email Settings",
-        onClick: () => {
-          NiceModal.show(EmailSettingsDialog, {
-            title,
-            enrollment: resource.data,
-          })
-        },
-      },
-      {
-        className: "dashboard-card-menu-item",
-        key: "unenroll",
-        label: "Unenroll",
-        onClick: () => {
-          NiceModal.show(UnenrollDialog, { title, enrollment: resource.data })
-        },
-      },
-    )
-
-    const receiptMenuItem = getReceiptMenuItem(
-      resource.data.enrollment_mode,
-      `/orders/receipt/by-run/${resource.data.run.id}/`,
-    )
-    if (receiptMenuItem) courseMenuItems.push(receiptMenuItem)
-
-    menuItems.push(...courseMenuItems)
-  }
-  return [...menuItems, ...additionalItems]
-}
-
-const getTitle = (
-  resource: DashboardResource,
-  selectedCourseRun?: BaseCourseRun | CourseRunV2 | null,
-): string => {
-  if (resource.type === DashboardType.Course) {
-    return selectedCourseRun?.title ?? resource.data.title
-  }
-  if (resource.type === DashboardType.CourseRunEnrollment) {
-    return resource.data.run.title
-  }
-  return resource.data.program.title
-}
-
-const getDashboardEnrollmentStatus = (
-  resource: DashboardResource,
-): EnrollmentStatus => {
-  const hasValidCertificate =
-    resource.type !== DashboardType.Course && !!resource.data.certificate?.uuid
-
-  if (resource.type === DashboardType.Course) {
-    return EnrollmentStatus.NotEnrolled
-  }
-
-  if (resource.type === DashboardType.CourseRunEnrollment) {
-    return hasValidCertificate
-      ? EnrollmentStatus.Completed
-      : getEnrollmentStatus(resource.data)
-  }
-
-  return hasValidCertificate
-    ? EnrollmentStatus.Completed
-    : EnrollmentStatus.Enrolled
-}
-
-const getDefaultNoun = (resource: DashboardResource): string => {
-  return resource.type === DashboardType.ProgramEnrollment
-    ? "Program"
-    : "Course"
-}
-
-const useEnrollmentHandler = () => {
-  const mitxOnlineUser = useQuery(mitxUserQueries.me())
-  const createB2bEnrollment = useCreateB2bEnrollment()
-  const createEnrollment = useCreateEnrollment()
-  const createVerifiedProgramEnrollment = useCreateVerifiedProgramEnrollment()
-  const replaceBasketItem = useReplaceBasketItem()
-
-  const enroll = React.useCallback(
-    ({
-      course,
-      readableId,
-      href,
-      selectedCoursewareUrl,
-      isB2B,
-      isVerifiedProgram,
-      programCoursewareId,
-      b2bProgramId,
-    }: {
-      course: CourseWithCourseRunsSerializerV2
-      readableId?: string
-      href?: string
-      selectedCoursewareUrl?: string
-      isB2B?: boolean
-      isVerifiedProgram?: boolean
-      programCoursewareId?: string
-      b2bProgramId?: string
-    }) => {
-      if (isB2B) {
-        if (!readableId) {
-          console.warn("Cannot enroll in B2B course: missing required data", {
-            readableId,
-            href,
-          })
-          return
-        }
-        const matchedRun = (course.courseruns ?? []).find(
-          (run) => run.courseware_id === readableId,
-        )
-        const destinationUrl = href ?? matchedRun?.courseware_url
-        if (!destinationUrl) {
-          console.warn("Cannot enroll in B2B course: missing destination URL", {
-            readableId,
-            href,
-          })
-          return
-        }
-        const userCountry = mitxOnlineUser.data?.legal_address?.country
-        const userYearOfBirth = mitxOnlineUser.data?.user_profile?.year_of_birth
-        const showJustInTimeDialog = !userCountry || !userYearOfBirth
-
-        if (showJustInTimeDialog) {
-          NiceModal.show(JustInTimeDialog, {
-            href: destinationUrl,
-            readableId,
-            programId: b2bProgramId,
-          })
-        } else {
-          createB2bEnrollment.mutate(
-            { readable_id: readableId, program_id: b2bProgramId },
-            {
-              onSuccess: () => {
-                window.location.href = destinationUrl
-              },
-            },
-          )
-        }
-      } else if (isVerifiedProgram && programCoursewareId && readableId) {
-        if (!href) {
-          console.warn(
-            "Cannot enroll in verified program course: missing href",
-            { href },
-          )
-          return
-        }
-        const verifiedDestination =
-          selectedCoursewareUrl ??
-          (course.courseruns ?? []).find(
-            (run) => run.courseware_id === readableId,
-          )?.courseware_url ??
-          href
-        createVerifiedProgramEnrollment.mutate(
-          { courserun_id: readableId, request_body: [programCoursewareId] },
-          {
-            onSuccess: () => {
-              window.location.href = verifiedDestination ?? href
-            },
-          },
-        )
-      } else {
-        const enrollmentAction = getCourseEnrollmentAction(course)
-
-        if (enrollmentAction.type === "audit") {
-          createEnrollment.mutate(
-            { run_id: enrollmentAction.run.id },
-            {
-              onSuccess: () => {
-                trackCourseEnrolled(course.title)
-                const destination =
-                  selectedCoursewareUrl ??
-                  enrollmentAction.run.courseware_url ??
-                  href
-                if (destination) {
-                  window.location.href = destination
-                }
-              },
-            },
-          )
-          return
-        }
-
-        if (enrollmentAction.type === "checkout") {
-          replaceBasketItem.mutate(enrollmentAction.product.id)
-          return
-        }
-
-        const onCourseEnroll = (run: CourseRunV2) => {
-          window.location.href = run.courseware_url!
-        }
-        NiceModal.show(CourseEnrollmentDialog, { course, onCourseEnroll })
-      }
-    },
-    [
-      mitxOnlineUser.data?.legal_address?.country,
-      mitxOnlineUser.data?.user_profile?.year_of_birth,
-      createB2bEnrollment,
-      createEnrollment,
-      createVerifiedProgramEnrollment,
-      replaceBasketItem,
-    ],
-  )
-
-  return {
-    enroll,
-    isPending:
-      createB2bEnrollment.isPending ||
-      createEnrollment.isPending ||
-      createVerifiedProgramEnrollment.isPending ||
-      replaceBasketItem.isPending,
-    mitxOnlineUser: mitxOnlineUser.data,
-  }
-}
-
-type CoursewareButtonProps = {
-  startDate?: string | null
-  endDate?: string | null
-  enrollmentStatus: EnrollmentStatus
-  href?: string | null
-  disabled?: boolean
-  className?: string
-  noun: string
-  isProgram?: boolean
-  isPending?: boolean
-  isStaff?: boolean
-  "data-testid"?: string
-  onClick?: React.MouseEventHandler<HTMLButtonElement>
-}
-
-const getCoursewareTextAndIcon = ({
-  endDate,
-  enrollmentStatus,
-  noun,
-  isProgram,
-}: {
-  endDate?: string | null
-  enrollmentStatus: EnrollmentStatus
-  noun: string
-  isProgram?: boolean
-}) => {
-  if (enrollmentStatus === EnrollmentStatus.NotEnrolled) {
-    return {
-      text: `Start ${noun}`,
-      endIcon: null,
-    }
-  }
-  if (
-    (endDate && isInPast(endDate)) ||
-    enrollmentStatus === EnrollmentStatus.Completed
-  ) {
-    return {
-      text: `View ${noun}`,
-      endIcon: null,
-    }
-  }
-  // Programs show "View Program" when enrolled, courses show "Continue"
-  if (isProgram && enrollmentStatus === EnrollmentStatus.Enrolled) {
-    return {
-      text: `View ${noun}`,
-      endIcon: null,
-    }
-  }
-  return {
-    text: "Continue",
-    endIcon: <RiArrowRightLine />,
-  }
-}
-
-const CoursewareButton = styled(
-  ({
-    startDate,
-    endDate,
-    enrollmentStatus,
-    href,
-    disabled,
-    className,
-    noun,
-    isProgram,
-    onClick,
-    isPending,
-    isStaff,
-    ...others
-  }: CoursewareButtonProps) => {
-    const coursewareText = getCoursewareTextAndIcon({
-      endDate,
-      noun,
-      enrollmentStatus,
-      isProgram,
-    })
-    const hasStarted = startDate && isInPast(startDate)
-    const hasEnrolled = enrollmentStatus !== EnrollmentStatus.NotEnrolled
-
-    // Programs or enrolled courses with started runs: show link
-    // Staff can access courseware even before the course has started
-    if (
-      (isProgram || hasEnrolled) &&
-      (hasStarted || !startDate || isStaff) &&
-      href
-    ) {
-      return (
-        <ButtonLink
-          size="small"
-          variant="primary"
-          endIcon={coursewareText.endIcon}
-          href={href}
-          className={className}
-          {...others}
-        >
-          {coursewareText.text}
-        </ButtonLink>
-      )
-    }
-
-    // Determine if button should be disabled
-    // Staff can access courseware even before the course has started
-    const isDisabled = Boolean(
-      disabled ||
-        (!hasEnrolled && !onClick) || // Not enrolled and no click handler
-        (hasEnrolled && !href && !onClick) || // Enrolled but no action available
-        (hasEnrolled && !!startDate && !hasStarted && !isStaff), // Enrolled but course hasn't started yet
-    )
-
-    return (
-      <Button
-        size="small"
-        variant="primary"
-        className={className}
-        onClick={onClick}
-        disabled={isDisabled}
-        endIcon={
-          isPending ? (
-            <LoadingSpinner color="inherit" loading={isPending} size={16} />
-          ) : (
-            coursewareText.endIcon
-          )
-        }
-        {...others}
-      >
-        {coursewareText.text}
-      </Button>
-    )
-  },
-)({ width: "124px" })
-
-const formatUpgradeTime = (daysFloat: number) => {
-  if (daysFloat < 0) return ""
-  const days = Math.floor(daysFloat)
-  if (days > 1) {
-    return `${days} days remaining`
-  } else if (days === 1) {
-    return `${days} day remaining`
-  }
-  return "Less than a day remaining"
-}
-
-const SubtitleLinkRoot = styled.div(({ theme }) => ({
-  display: "flex",
-  alignItems: "center",
-  gap: "8px",
-  flex: 1,
-  color: theme.custom.colors.darkGray2,
-  ...theme.typography.subtitle3,
-}))
-const SubtitleLink = styled(NextLink)(({ theme }) => ({
-  ...theme.typography.subtitle3,
-  color: theme.custom.colors.mitRed,
-  display: "flex",
-  alignItems: "center",
-  gap: "4px",
-  ":hover": {
-    textDecoration: "underline",
-  },
-}))
-
-const UpgradeBanner: React.FC<
-  {
-    canUpgrade: boolean
-    certificateUpgradeDeadline?: string | null
-    certificateUpgradePrice?: string | null
-    productId?: number | null
-    onError?: (error: Error) => void
-  } & React.HTMLAttributes<HTMLDivElement>
-> = ({
-  canUpgrade,
-  certificateUpgradeDeadline,
-  certificateUpgradePrice,
-  productId,
-  onError,
-  ...others
-}) => {
-  const replaceBasketItem = useReplaceBasketItem()
-
-  const handleUpgradeClick = async (e: React.MouseEvent<HTMLAnchorElement>) => {
-    e.preventDefault()
-    if (!productId) return
-
-    try {
-      await replaceBasketItem.mutateAsync(productId)
-    } catch (error) {
-      onError?.(error as Error)
-    }
-  }
-
-  if (!canUpgrade || !certificateUpgradePrice || !productId) {
-    return null
-  }
-
-  // If deadline is provided, check it hasn't passed
-  if (certificateUpgradeDeadline && isInPast(certificateUpgradeDeadline)) {
-    return null
-  }
-
-  const formattedPrice = `$${certificateUpgradePrice}`
-  const calendarDays = certificateUpgradeDeadline
-    ? calendarDaysUntil(certificateUpgradeDeadline)
-    : null
-
-  return (
-    <SubtitleLinkRoot {...others}>
-      <SubtitleLink href="#" onClick={handleUpgradeClick}>
-        <RiAddLine size="16px" />
-        {`Add a certificate for ${formattedPrice}`}
-      </SubtitleLink>
-      {calendarDays !== null && (
-        <NoSSR>
-          {/* This uses local time. */}
-          {formatUpgradeTime(calendarDays)}
-        </NoSSR>
-      )}
-    </SubtitleLinkRoot>
-  )
-}
-
-const CountdownRoot = styled.div(({ theme }) => ({
-  width: "100%",
-  paddingRight: "32px",
-  display: "flex",
-  justifyContent: "center",
-  alignSelf: "end",
-  [theme.breakpoints.down("md")]: {
-    marginRight: "0px",
-    justifyContent: "flex-start",
-  },
-}))
-const CourseStartCountdown: React.FC<{
-  startDate: string
-  className?: string
-}> = ({ startDate, className }) => {
-  const calendarDays = calendarDaysUntil(startDate)
-
-  let value
-  if (calendarDays === null || calendarDays < 0) return null
-  if (calendarDays === 0) {
-    value = "Starts Today"
-  } else if (calendarDays === 1) {
-    value = "Starts Tomorrow"
-  } else {
-    value = `Starts in ${calendarDays} days`
-  }
-  return (
-    <CountdownRoot>
-      <Link
-        color="black"
-        size="small"
-        className={className}
-        onClick={console.log}
-      >
-        {value}
-      </Link>
-    </CountdownRoot>
-  )
-}
-
-type DashboardCardProps = {
-  resource: DashboardResource
-  showNotComplete?: boolean
-  offerUpgrade?: boolean
-  noun?: string
-  contextMenuItems?: SimpleMenuItem[]
-  isLoading?: boolean
-  buttonClick?: React.MouseEventHandler<HTMLButtonElement>
-  Component?: React.ElementType
-  className?: string
-  variant?: "default" | "stacked"
-  contractId?: number
-  programEnrollment?: V3UserProgramEnrollment
-  parentProgramReadableIds?: string[]
-  onUpgradeError?: (error: string) => void
-  selectedCourseRun?: BaseCourseRun | CourseRunV2 | null
-  uiLanguageCode?: string
-}
-
-const DashboardCard: React.FC<DashboardCardProps> = ({
-  resource,
-  showNotComplete = true,
-  offerUpgrade = true,
-  noun,
-  contextMenuItems = [],
-  isLoading = false,
-  buttonClick,
-  Component,
-  className,
-  variant = "default",
-  contractId,
-  programEnrollment,
-  parentProgramReadableIds,
-  onUpgradeError,
-  selectedCourseRun,
-  uiLanguageCode: _uiLanguageCode = "en",
-}) => {
-  const enrollment = useEnrollmentHandler()
-  const mitxOnlineUser = enrollment.mitxOnlineUser
-  const useProductPages = useFeatureFlagEnabled(
-    FeatureFlags.MitxOnlineProductPages,
-  )
-
-  const title = getTitle(resource, selectedCourseRun)
-  const courseRun =
-    resource.type === DashboardType.Course
-      ? (selectedCourseRun ??
-        getBestRun(resource.data, { enrollableOnly: true, contractId }))
-      : undefined
-  const enrollmentRun =
-    resource.type === DashboardType.CourseRunEnrollment
-      ? resource.data.run
-      : undefined
-  const enrollmentStatus = getDashboardEnrollmentStatus(resource)
-  const certificateLink = getCertificateLink(resource)
-  const displayNoun = noun ?? getDefaultNoun(resource)
-
-  const isCourse = resource.type === DashboardType.Course
-  const isCourseRunEnrollment =
-    resource.type === DashboardType.CourseRunEnrollment
-  const isProgramEnrollment = resource.type === DashboardType.ProgramEnrollment
-  const isAnyCourse = isCourse || isCourseRunEnrollment
-
-  const coursewareUrl = isCourse
-    ? courseRun?.courseware_url
-    : enrollmentRun?.courseware_url
-  const b2bContractId =
-    // courseRun may be a BaseCourseRun (from the variant-runs endpoint, which
-    // uses a lighter serializer that omits b2b_contract).  The cast is safe:
-    // if it IS a BaseCourseRun, b2b_contract is simply undefined at runtime
-    // and the fallback chain below picks up contractId instead.
-    (courseRun as CourseRunV2 | undefined)?.b2b_contract ??
-    (resource.type === DashboardType.CourseRunEnrollment
-      ? resource.data.b2b_contract_id
-      : undefined) ??
-    contractId
-  // TODO: Replace this inferred contract-page check once include_in_learn_catalog is available in v3.
-  const isContractPageResource = Boolean(b2bContractId)
-
-  const hasEnrollableRuns = isCourse
-    ? (courseRun?.is_enrollable ?? false)
-    : true
-
-  const disableEnrollment = isCourse && !hasEnrollableRuns
-
-  const readableId = isCourse
-    ? courseRun?.courseware_id
-    : isCourseRunEnrollment
-      ? resource.data.run.courseware_id
-      : isProgramEnrollment
-        ? resource.data.program.readable_id
-        : undefined
-
-  const canUpgrade =
-    isCourseRunEnrollment &&
-    !isVerifiedEnrollmentMode(resource.data.enrollment_mode) &&
-    (enrollmentRun?.is_upgradable ?? false) &&
-    (enrollmentRun?.upgrade_product_is_active ?? false)
-
-  // Handle enrollment click for courses
-  const handleEnrollmentClick = React.useCallback(() => {
-    if (isCourse) {
-      const isVerifiedProgramEnrollment = isVerifiedEnrollmentMode(
-        programEnrollment?.enrollment_mode,
-      )
-
-      enrollment.enroll({
-        course: resource.data,
-        readableId: readableId,
-        href: coursewareUrl ?? undefined,
-        selectedCoursewareUrl: coursewareUrl ?? undefined,
-        isB2B: !!b2bContractId,
-        isVerifiedProgram: isVerifiedProgramEnrollment,
-        programCoursewareId: programEnrollment?.program.readable_id,
-        b2bProgramId:
-          parentProgramReadableIds?.[0] ??
-          programEnrollment?.program.readable_id,
-      })
-    }
-  }, [
-    isCourse,
-    resource,
-    readableId,
-    coursewareUrl,
-    b2bContractId,
-    enrollment,
-    programEnrollment?.enrollment_mode,
-    programEnrollment?.program.readable_id,
-    parentProgramReadableIds,
-  ])
-
-  // Determine title behavior (link vs clickable text vs plain text)
-  const titleHref = isCourseRunEnrollment
-    ? coursewareUrl
-    : isProgramEnrollment
-      ? programView(resource.data.program.id)
-      : undefined
-
-  const titleClick: React.MouseEventHandler | undefined = isCourse
-    ? (e) => {
-        e.preventDefault()
-        handleEnrollmentClick()
-      }
-    : undefined
-
-  // Button onClick handler
-  const coursewareButtonClick:
-    | React.MouseEventHandler<HTMLButtonElement>
-    | undefined = isCourse ? handleEnrollmentClick : buttonClick
-
-  // Build title section
-  const titleSection = isLoading ? (
-    <>
-      <Skeleton variant="text" width="95%" height={16} />
-      <Skeleton variant="text" width={120} height={16} />
-      <Skeleton variant="text" width={120} height={16} />
-    </>
-  ) : (
-    <>
-      {titleHref ? (
-        <TitleHeading>
-          <TitleLink
-            size="medium"
-            color="black"
-            href={titleHref}
-            onClick={titleClick}
-          >
-            {title}
-          </TitleLink>
-        </TitleHeading>
-      ) : titleClick ? (
-        <TitleText clickable onClick={titleClick}>
-          {title}
-        </TitleText>
-      ) : (
-        <TitleText>{title}</TitleText>
-      )}
-      {certificateLink ? (
-        <SubtitleLink href={certificateLink}>
-          <RiAwardLine size="16px" />
-          View Certificate
-        </SubtitleLink>
-      ) : null}
-      {isCourseRunEnrollment &&
-      !isVerifiedEnrollmentMode(resource.data.enrollment_mode) &&
-      offerUpgrade ? (
-        <UpgradeBanner
-          data-testid="upgrade-root"
-          canUpgrade={canUpgrade}
-          certificateUpgradeDeadline={enrollmentRun?.upgrade_deadline}
-          certificateUpgradePrice={enrollmentRun?.upgrade_product_price}
-          productId={enrollmentRun?.upgrade_product_id}
-          onError={() => {
-            onUpgradeError?.(
-              "There was a problem adding the certificate to your cart.",
-            )
-          }}
-        />
-      ) : null}
-    </>
-  )
-
-  // Build button section
-  const buttonSection = isLoading ? (
-    <Skeleton variant="rectangular" width={120} height={32} />
-  ) : isAnyCourse ? (
-    <>
-      <EnrollmentStatusIndicator
-        status={enrollmentStatus}
-        showNotComplete={showNotComplete}
-      />
-      <CoursewareButton
-        data-testid="courseware-button"
-        startDate={isCourse ? courseRun?.start_date : enrollmentRun?.start_date}
-        enrollmentStatus={enrollmentStatus}
-        href={coursewareUrl ?? undefined}
-        endDate={isCourse ? courseRun?.end_date : enrollmentRun?.end_date}
-        noun={displayNoun}
-        isProgram={false}
-        disabled={disableEnrollment}
-        isPending={enrollment.isPending}
-        isStaff={mitxOnlineUser?.is_staff}
-        onClick={coursewareButtonClick}
-      />
-    </>
-  ) : isProgramEnrollment &&
-    resource.type === DashboardType.ProgramEnrollment ? (
-    <CoursewareButton
-      noun={displayNoun}
-      isProgram={true}
-      enrollmentStatus={enrollmentStatus}
-      href={programView(resource.data.program.id)}
-    />
-  ) : null
-
-  // Build start date section
-  const startDateSection = isLoading ? (
-    <Skeleton variant="text" width={100} height={24} />
-  ) : isAnyCourse &&
-    (isCourse ? courseRun?.start_date : enrollmentRun?.start_date) ? (
-    <CourseStartCountdown
-      startDate={
-        isCourse
-          ? (courseRun?.start_date as string)
-          : (enrollmentRun?.start_date as string)
-      }
-    />
-  ) : null
-
-  // Build context menu
-  const menuItems = getContextMenuItems(
-    title,
-    resource,
-    useProductPages ?? false,
-    contextMenuItems,
-    isContractPageResource,
-  )
-
-  const contextMenu = isLoading ? (
-    <Skeleton variant="rectangular" width={12} height={24} />
-  ) : (
-    <SimpleMenu
-      items={menuItems}
-      trigger={
-        <MenuButton
-          size="small"
-          variant="text"
-          aria-label="More options"
-          status={enrollmentStatus}
-          hidden={menuItems.length === 0}
-        >
-          <RiMore2Line />
-        </MenuButton>
-      }
-    />
-  )
-
-  return (
-    <>
-      <CardRoot
-        screenSize="desktop"
-        data-testid="enrollment-card-desktop"
-        as={Component}
-        className={className}
-        variant={variant}
-      >
-        <Stack justifyContent="start" alignItems="stretch" gap="8px" flex={1}>
-          {titleSection}
-        </Stack>
-        <Stack gap="8px">
-          <Stack direction="row" gap="8px" alignItems="center">
-            {buttonSection}
-            {contextMenu}
-          </Stack>
-          {startDateSection}
-        </Stack>
-      </CardRoot>
-
-      <CardRoot
-        screenSize="mobile"
-        data-testid="enrollment-card-mobile"
-        as={Component}
-        className={className}
-        variant={variant}
-      >
-        <Stack
-          direction="row"
-          justifyContent="space-between"
-          alignItems="stretch"
-          flex={1}
-          width="100%"
-        >
-          <Stack direction="column" gap="8px" flex={1}>
-            {titleSection}
-          </Stack>
-          {contextMenu}
-        </Stack>
-        <Stack
-          direction="row"
-          alignItems="center"
-          justifyContent="end"
-          width="100%"
-        >
-          {startDateSection}
-          <Stack direction="row" gap="8px" alignItems="center">
-            {buttonSection}
-          </Stack>
-        </Stack>
-      </CardRoot>
-    </>
-  )
-}
-
-export { DashboardCard, CardRoot as DashboardCardRoot, getContextMenuItems }
