--- frontends/main/src/app-pages/DashboardPage/CoursewareDisplay/ModuleCard.test.tsx
@@ -1,65 +1,0 @@
-import * as mitxonline from "api/mitxonline-test-utils"
-import { EnrollmentModeEnum } from "@mitodl/mitxonline-api-axios/v2"
-import { mitxonlineLegacyUrl } from "@/common/mitxonline"
-import { DashboardType, getContextMenuItems } from "./ModuleCard"
-
-describe("ModuleCard context menu receipt item", () => {
-  test("shows Receipt item for verified enrollment", () => {
-    const course = mitxonline.factories.courses.course()
-    const run = mitxonline.factories.courses.courseRun({ id: 42 })
-    const enrollment = mitxonline.factories.enrollment.courseEnrollment({
-      enrollment_mode: EnrollmentModeEnum.Verified,
-      run: { ...run, course },
-    })
-
-    const items = getContextMenuItems("Test Course", {
-      type: DashboardType.CourseRunEnrollment,
-      data: enrollment,
-    })
-
-    expect(items.some((item) => item.label === "Receipt")).toBe(true)
-  })
-
-  test("does not show Receipt item for audit enrollment", () => {
-    const course = mitxonline.factories.courses.course()
-    const run = mitxonline.factories.courses.courseRun({ id: 42 })
-    const enrollment = mitxonline.factories.enrollment.courseEnrollment({
-      enrollment_mode: EnrollmentModeEnum.Audit,
-      run: { ...run, course },
-    })
-
-    const items = getContextMenuItems("Test Course", {
-      type: DashboardType.CourseRunEnrollment,
-      data: enrollment,
-    })
-
-    expect(items.some((item) => item.label === "Receipt")).toBe(false)
-  })
-
-  test("Receipt item opens the expected MITx Online URL", () => {
-    const course = mitxonline.factories.courses.course()
-    const run = mitxonline.factories.courses.courseRun({ id: 42 })
-    const enrollment = mitxonline.factories.enrollment.courseEnrollment({
-      enrollment_mode: EnrollmentModeEnum.Verified,
-      run: { ...run, course },
-    })
-    const windowOpenSpy = jest
-      .spyOn(window, "open")
-      .mockImplementation(() => null)
-
-    const items = getContextMenuItems("Test Course", {
-      type: DashboardType.CourseRunEnrollment,
-      data: enrollment,
-    })
-    const receiptItem = items.find((item) => item.label === "Receipt")
-
-    receiptItem?.onClick?.()
-
-    expect(windowOpenSpy).toHaveBeenCalledWith(
-      mitxonlineLegacyUrl("/orders/receipt/by-run/42/"),
-      "_blank",
-      "noopener,noreferrer",
-    )
-    windowOpenSpy.mockRestore()
-  })
-})
