+++ frontends/main/src/app-pages/DashboardPage/CoursewareDisplay/OrganizationCards.tsx
@@ -3,7 +3,7 @@ import Image from "next/image"
 import graduateLogo from "@/public/images/dashboard/graduate.png"
 import { Stack, Typography, styled, theme, Link } from "ol-components"
 import { useQuery } from "@tanstack/react-query"
-import { CoursewareCardRoot } from "./CoursewareCard"
+import { CardRoot } from "./CardShared"
 import { mitxUserQueries } from "api/mitxonline-hooks/user"
 import { ButtonLink } from "@mitodl/smoot-design"
 import { contractView } from "@/common/urls"
@@ -43,7 +43,7 @@ const ContractCard = styled.div({
   },
 })
 
-const CardRootStyled = styled(CoursewareCardRoot)({
+const CardRootStyled = styled(CardRoot)({
   display: "flex",
   flexDirection: "column",
   padding: 0,
